feat: add test Main and Profile
