+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "start": "webpack serve --config webpack.config.js --open",
     "build-prod": "set NODE_ENV=production&&webpack",
-    "test": "jest --config jest.config.js",
-    "test-watch": "jest --config jest.config.js --watchAll",
+    "test": "jest --config jest.config.js --setupFiles dotenv/config",
+    "test-watch": "jest --config jest.config.js --setupFiles dotenv/config --watchAll",
     "clear": "rd /s /q dist",
     "lint": "eslint ./src --color",
     "lint:fix": "eslint ./src --fix --color",
